requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
yt-dlp==2023.9.24
youtube-transcript-api==0.6.1
groq==0.4.1
requests==2.31.0
urllib3==2.0.7